remove directory structure tool
...in favor of a `directoriesOnly` argument to `getProjectTree`.
