+++ extensions/positron-assistant/src/types.ts
@@ -16,7 +16,6 @@ export enum PositronAssistantToolName {
 	InspectVariables = 'inspectVariables',
 	SelectionEdit = 'selectionEdit',
 	ProjectTree = 'getProjectTree',
-	DirectoryStructure = 'getDirectoryStructure',
 	GetChangedFiles = 'getChangedFiles',
 	DocumentCreate = 'documentCreate',
 	TextSearch = 'positron_findTextInProject_internal',
